vgpt_requirements.txt
absl-py
asn1crypto
cachetools
certifi
cffi
chardet
click
cryptography
cycler
cymem
Cython
cytoolz
dataclasses
sp
dill
#en-core-web-sm
filelock
future
google-auth
google-auth-oauthlib
grpcio
spacy
h5py
idna
joblib
kiwisolver
Markdown
mkl-fft
mkl-random
mkl-service
msgpack
msgpack-numpy
multiprocess
murmurhash
oauthlib
packaging
pathlib
pathos
Pillow
plac
pox
ppft
preshed
protobuf
pyasn1
pyasn1-modules
pycocotools
pycparser
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytorch-lightning
pytz
regex
requests
requests-oauthlib
rsa
sacremoses
sentencepiece
six
scikit-image
scikit-learn
scipy
tensorboard
tensorboard-plugin-wit
termcolor
thinc
tokenizers
toolz
tqdm
transformers==4.12.5
ujson
urllib3
Werkzeug
wrapt
PyYAML
pandas
numpy
setuptools==59.5.0
opencv-python
matplotlib
huggingface-hub
torchmetrics
SimpleITK
SimpleITK-SimpleElastix
pydicom
nibabel
h5py
seaborn
xlrd
nltk
rouge
openpyxl
albumentations
timm
scispacy
monai
MedPy
dicom2nifti
torchsummary
ipykernel
pycocotools-windows
tensorflow-gpu
PySoundFile
selenium
pandas
numpy
einops
#conda install -c conda-forge openjdk=11